Moved static array from function to global scope
gcc does not allow a static array with unkown size at compile time.
The size of the array is after precompilation ((unsigned)(22050 / 18.2)).
A cast is not considered to be a constant expression.
Instead of putting the variable to global scope (this is what
this commit does) one could either remove the static modifier
(unkown effect?) or use an integer instead of a float in the
expression to avoid the casting.