Fix crash when saving plans on empty slots
On Linux, the game crashed when trying to save plans in slots where the
save file did not yet exist, when the datadisk directory is not written
in all lowercase.

This has been fixed. In case the save file cannot be found, only
perform a case-insensitive lookup of the directory component beforehand.